requirements.txt
# Financial AI Assistant - Python Dependencies

# Core dependencies
requests>=2.31.0
pandas>=2.0.0
openai>=1.0.0
elasticsearch>=8.0.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0

# Optional: For advanced analytics
numpy>=1.24.0
scipy>=1.11.0

# Optional: For visualization (if needed)
matplotlib>=3.7.0
plotly>=5.14.0

# Optional: For web interface (future enhancement)
flask>=3.0.0
flask-cors>=4.0.0
